pothole-ai-system/ai/requirements.txt
openai
python-dotenv
requests
resend
moorcheh-sdk